requirements.txt
pytorch-lightning==1.4.7
torch==1.9.0
torchvision==0.10.0
opencv-python==4.5.3.56
Pillow==8.3.2
click==8.0.1
seaborn==0.11.2
